requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.9.10
django-filter==2.4.0
python-decouple==3.8
setuptools==78.1.0